requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
opencv-python==4.7.0.72
moviepy==1.0.3
transformers==4.26.1
torch==2.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.3
openai-whisper
Pillow==9.5.0
numpy==1.23.5
huggingface_hub==0.11.1
